Narrow six returns that ty 0.0.69 calls unsound (#25)
The pending dependency bump takes ty from 0.0.65 to 0.0.69, which adds
`unsound-return-statement`: a value inferred as Any (or a union containing it)
returned from a function that declares a concrete type. Six sites trip it and
the bump's CI fails on all three Python versions because of them.

None is in `src/`. Five come from arithmetic over values the checker cannot see
into, and one from `tomllib`, which is Any by construction. The fix is the
narrowing the diagnostic asks for rather than a rule exemption: `float(...)` and
`int(...)` assert at runtime what the annotation already promised, and the two
`_agreement` calls get annotated locals because the script they come from is
imported dynamically, so every attribute off it is Any.

Verified under both versions -- 0.0.65 as pinned on main and 0.0.69 as the bump
would install -- with `uv run ty check`, `ruff format --check`, and `ruff check`
run exactly as ci.yml runs them, plus the tests for the three test files touched.
No lockfile change here: pinning the new ty is the bump's job, not this one's.

+++ tests/test_escapes.py
@@ -173,7 +173,7 @@ def _relative_luminance(hex_colour: str) -> float:
         value = int(raw[offset : offset + 2], 16) / 255
         channels.append(value / 12.92 if value <= 0.03928 else ((value + 0.055) / 1.055) ** 2.4)
     red, green, blue = channels
-    return 0.2126 * red + 0.7152 * green + 0.0722 * blue
+    return float(0.2126 * red + 0.7152 * green + 0.0722 * blue)
 
 
 def _contrast(foreground: str, background: str) -> float:
